Add files via upload
Support Vector Machine (SVM) Classifier is the right model due to high accuracy, precision and recall score. One reason why SVM Classifier compared to Random Forest Classifier showed an improved performance was because of the presence of outliers. Here is the accuracy for the different approaches
* Support Vector Machine  – 83%
 
* Random Forest Classifier  – 79%
 
* KMean  – 67%
 
As mentioned before, since Random Forest is not a distance-based algorithm, it is not much affected by outliers, whereas distance-based algorithm such as KMean showed a lower performance.

Based on the feature importance :

Glucose is the most important factor in determining the onset of diabetes followed by BMI and Age. Other factors such as Diabetes Pedigree Function, Pregnancies, Blood Pressure, Skin Thickness and Insulin also contributes to the prediction.

As we can see, the results derived from Feature Importance makes sense as one of the first things that actually is monitored in high-risk patients is the Glucose level.
An increased BMI might also indicate a risk of developing Type II Diabetes. Normally, especially in case of Type II Diabetes, there is a high risk of developing as the age of a person increases (given other factors).